+++ openaq_api/openaq_api/exceptions.py
@@ -1,0 +1,15 @@
+
+from fastapi import (
+    HTTPException,
+    status,
+    )
+
+NOT_AUTHENTICATED_EXCEPTION = HTTPException(
+    status_code=status.HTTP_401_UNAUTHORIZED,
+    detail="Invalid credentials",
+)
+
+TOO_MANY_REQUESTS = HTTPException(
+    status_code=status.HTTP_429_TOO_MANY_REQUESTS,
+    detail="To many requests",
+)
